+++ packages/wrapper/web/src/localeStrategy/localeStrategyProvider.js
@@ -1,7 +1,12 @@
 export default ({ locale = 'en' } = { locale: 'en' }) => (innerProps = {}) => {
-  if (locale === 'en') {
+  const mergedLocale = innerProps.locale || locale;
+  if (mergedLocale === 'en') {
     return innerProps;
   }
-  const localeResource = require(`wix-rich-content-common/dist/statics/locale/messages_${locale}.json`); // eslint-disable-line
-  return { ...innerProps, localeResource };
+  try {
+    const localeResource = require(`wix-rich-content-common/dist/statics/locale/messages_${mergedLocale}.json`); // eslint-disable-line
+    return { locale: mergedLocale, localeResource, ...innerProps };
+  } catch (err) {
+    throw new Error(`error while loading locale ${locale}`, err);
+  }
 };