+++ CHANGELOG.md
@@ -19,6 +19,8 @@
 - `wrapper`
   - [#714](https://github.com/wix-incubator/rich-content/pull/714) locale strategy added
 ### :bug: Bug Fix
+- `wrapper`
+  - [#719](https://github.com/wix-incubator/rich-content/pull/719) locale strategy fixes
 - `image`
   - [#712](https://github.com/wix-incubator/rich-content/pull/712) old image client data support #2
 ### :house: Internal